Separado backend y frontend - eliminado cartaController y precios en backend son BigNumber

+++ Backend/src/main/java/com/example/cartasIntercambio/controller/PublicacionController.java
@@ -15,6 +15,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -51,7 +52,7 @@ public class PublicacionController {
     }
 
     @GetMapping("/precio/{precio}") 
-    public ResponseEntity<List<PublicacionDto>> buscarPublicacionPorPrecioDeCarta(@PathVariable float precio) {
+    public ResponseEntity<List<PublicacionDto>> buscarPublicacionPorPrecioDeCarta(@PathVariable BigDecimal precio) {
         return new ResponseEntity<>(publicacionService.buscarPublicacionPorPrecio(precio), HttpStatus.OK);
     }
