+++ Backend/src/main/java/com/example/cartasIntercambio/model/Mercado/Demanda.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,6 +14,6 @@ import java.util.List;
 public class Demanda {
 
     private Carta cartaOfrecida;
-    private Float precio;
+    private BigDecimal precio;
     private List<Carta> cartasInteres;
 }
